+++ js/carousel.js
@@ -719,7 +719,7 @@
 
     // Return error if an unrecognized  method name is passed in
     } else {
-      jQuery.error(`Method ${methodOrOptions} does not exist on jQuery.collapsible`);
+      jQuery.error(`Method ${methodOrOptions} does not exist on jQuery.carousel`);
     }
   };
 }( cash ));